cache in run_classifier + various fixes to the examples
